+++ changelog.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented here. :)
 
+## MITM 0.0.2
+Implement core functionality. Released on Nov. 30, 2017
+
+### Added
+* Implement basic resource file replace functionality. (Only support *.js currently)
+* Use popup dialog to enable/disable extension.
+
 ## MITM 0.0.1
 Released on Nov. 25, 2017